requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
crispy-bootstrap5==2025.6
cryptography==46.0.3
dj-database-url==3.0.1
Django==5.2.7
django-anymail==13.1
django-cloudinary-storage==0.3.0
django-crispy-forms==2.4
django-filter==25.2
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.11
pycparser==2.23
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
stripe==13.2.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
